tech-pulse-engine/requirements.txt
SQLAlchemy>=2.0,<3.0
psycopg[binary]>=3.2,<4.0
feedparser>=6.0,<7.0
requests>=2.32,<3.0
beautifulsoup4>=4.12,<5.0
APScheduler>=3.10,<4.0
python-dotenv>=1.0,<2.0
PyYAML>=6.0,<7.0
fastapi>=0.115,<1.0
uvicorn>=0.34,<1.0
Jinja2>=3.1,<4.0
